+++ R/check-mac.R
@@ -5,23 +5,21 @@
 #' link to the page with the check results.
 #'
 #' @template devtools
-#' @inheritParams pkgbuild::build
+#' @inheritParams check_win
 #' @param dep_pkgs Additional custom dependencies to install prior to checking the package.
-#' @param quiet If `TRUE`, suppresses output.
-#' @param ... Additional arguments passed to [pkgbuild::build()].
 #' @family build functions
 #' @return The url with the check results (invisibly)
 #' @export
-check_mac_release <- function(pkg = ".", dep_pkgs = ".", args = NULL, manual = TRUE, quiet = FALSE, ...) {
+check_mac_release <- function(pkg = ".", dep_pkgs = character(), args = NULL, manual = TRUE, quiet = FALSE, ...) {
   check_dots_used(action = getOption("devtools.ellipsis_action", rlang::warn))
 
   pkg <- as.package(pkg)
 
   if (!quiet) {
-    cli::cli_alert_info(
+    cli::cli_inform(c(
       "Building macOS version of {.pkg {pkg$package}} ({pkg$version})",
-      "with https://mac.r-project.org/macbuilder/submit.html."
-    )
+      i = "Using https://mac.r-project.org/macbuilder/submit.html."
+    ))
   }
 
   built_path <- pkgbuild::build(pkg$path, tempdir(),
@@ -64,9 +62,10 @@ check_mac_release <- function(pkg = ".", dep_pkgs = ".", args = NULL, manual = T
   if (!quiet) {
     time <- strftime(Sys.time() + 10 * 60, "%I:%M %p")
 
-    cli::cli_alert_success(
-      "[{Sys.Date()}] Check {.url {response_url}} for the results in 5-10 mins (~{time})."
-    )
+    cli::cat_rule(col = "cyan")
+    cli::cli_inform(c(
+      i = "Check {.url {response_url}} the results in 5-10 mins (~{time})."
+    ))
   }
 
   invisible(response_url)