+++ tests/testthat/test-install.R
@@ -2,7 +2,7 @@ library(mockery)
 
 local({
 
-pkg <- create_local_package()
+pkg <- fs::path_real(local_package_create())
 
 path2char <- function(x) {
   if (inherits(x, "fs_path")) {
@@ -16,7 +16,7 @@ expect_passes_args <- function(fn, stub, input_args = list(), expected_args) {
   mck <- mockery::mock(NULL)
   mockery::stub(fn, stub, mck)
 
-  suppressMessages(do.call(fn, input_args))
+  capture.output(suppressMessages(do.call(fn, input_args)))
 
   mockery::expect_called(mck, 1)
   mock_args <- mockery::mock_args(mck)[[1]]
@@ -110,4 +110,21 @@ test_that("install_dev_deps passes ellipsis args to remotes::install_deps", {
   )
 })
 
+})
+
+test_that("vignettes built on install", {
+  skip_on_cran()
+
+  if (!pkgbuild::has_latex()) {
+    skip("pdflatex not available")
+  }
+
+  pkg <- local_package_copy(test_path("testVignettesBuilt"))
+
+  withr::local_temp_libpaths()
+  install(pkg, reload = FALSE, quiet = TRUE, build_vignettes = TRUE)
+
+  vigs <- vignette(package = "testVignettesBuilt")$results
+  expect_equal(nrow(vigs), 1)
+  expect_equal(vigs[3], "new")
 })